distinguish between FileModelBundleValidator and AssetModelBundleValidator
