+++ core/src/main/java/ai/doc/tensorio/core/modelbundle/ModelBundleValidator.java
@@ -25,27 +25,20 @@ import android.content.Context;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.github.fge.jackson.JsonLoader;
 import com.github.fge.jsonschema.core.exceptions.ProcessingException;
-import com.github.fge.jsonschema.core.report.ProcessingReport;
 import com.github.fge.jsonschema.main.JsonSchema;
 import com.github.fge.jsonschema.main.JsonSchemaFactory;
 
-import org.json.JSONArray;
-import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.Arrays;
 import java.util.function.BiFunction;
 
 import ai.doc.tensorio.core.utilities.AndroidAssets;
-import ai.doc.tensorio.core.utilities.FileIO;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
-// TODO: Split into two classes, one for File, one for Asset
-
-public class ModelBundleValidator {
+public abstract class ModelBundleValidator {
 
     public static class ValidatorException extends Exception {
         public ValidatorException(@NonNull String message) {
@@ -57,66 +50,44 @@ public class ModelBundleValidator {
         }
     }
 
-    /** Source is an asset from a context or a file. Barf */
-
-    public enum Source {
-        Asset,
-        File
-    }
-
-    private Source source;
-
-    /** Used to validate model bundles in an assets directory */
-
-    private Context context;
-    private String filename;
-    private String jsonFilename;
-
-    /** Used to validate model bundle at any other File path */
-
-    private File file;
-    private File jsonFile;
-
     /**
-     * Initializes the Model Bundle Validator with an asset
+     * Creates and returns a new AssetModelBundleValidator for the bundle in the assets directory at path
      *
-     * Model validation is done with json-schema-validator library using the JSON Schema Internet draft Version 4.
-     * More info about the schema draft can be found at https://tools.ietf.org/html/draft-zyp-json-schema-04.
-     *
-     * @param context Context used to acquire the TFLite/model-schema.json and the filename
-     * @param filename The model bundle's filename relative to the assets directory
+     * @param context The application or activity context
+     * @param filename The filename for the model bundle that will be validated, including subdirectories
+     * @return A @see AssetModelBundleValidator
      */
 
-    public ModelBundleValidator(@NonNull Context context, @NonNull String filename) {
-        this.source = Source.Asset;
-        this.context = context;
-
-        this.filename = filename;
-        this.jsonFilename = filename + "/" + ModelBundle.TFMODEL_INFO_FILE;
-
-        this.file = null;
-        this.jsonFile = null;
+    public static ModelBundleValidator validatorWithAsset(@NonNull Context context, @NonNull String filename) {
+        return new AssetModelBundleValidator(context, filename);
     }
 
     /**
-     * Initializes the Model Bundle Validator with a file
+     * Creates and returns a new FileModelBundleValidator for the bundle at File
      *
-     * Model validation is done with json-schema-validator library using the JSON Schema Internet draft Version 4.
-     * More info about the schema draft can be found at https://tools.ietf.org/html/draft-zyp-json-schema-04.
-     *
-     * @param context Context used to acquire the TFLite/model-schema.json
-     * @param file Fully qualified File pointing to the model bundle
+     * @param context The application or activity context
+     * @param file The File to the model bundle
+     * @return A @see FileModelBundleValidator
      */
 
-    public ModelBundleValidator(@NonNull Context context, @NonNull File file) {
-        this.source = Source.File;
+    public static ModelBundleValidator validatorWithFile(@NonNull Context context, @NonNull File file) {
+        return new FileModelBundleValidator(context, file);
+    }
+
+    /**
+     * The application or activity context that contains the JSON schema
+     */
+
+    protected Context context;
+
+    /**
+     * Designated initializer although this is an abstract class so it just sets the context property
+     *
+     * @param context The application or activity context
+     */
+
+    public ModelBundleValidator(Context context) {
         this.context = context;
-
-        this.file = file;
-        this.jsonFile = new File(file, ModelBundle.TFMODEL_INFO_FILE);
-
-        this.filename = null;
-        this.jsonFilename = null;
     }
 
     /**
@@ -140,257 +111,11 @@ public class ModelBundleValidator {
      * @throws ValidatorException On any validation error. See the message for more details
      */
 
-    public void validate(@Nullable BiFunction<String, JSONObject, Boolean> customValidator) throws ValidatorException {
-        switch (source) {
-            case Asset:
-                validateAsset(customValidator);
-                break;
-            case File:
-                validateFile(customValidator);
-                break;
-        }
-    }
-
-    /** Validates a model bundle in the asset directory of a package */
-
-    private void validateAsset(@Nullable BiFunction<String, JSONObject, Boolean> customValidator) throws ValidatorException {
-        // Validate Path
-
-        try {
-            if (!Arrays.asList(context.getResources().getAssets().list("")).contains(filename)) {
-                throw new ValidatorException("No model bundle at found at provided path");
-            }
-        } catch (Exception e) {
-            throw new ValidatorException("No model bundle at provided path");
-        }
-
-        // Validate Bundle Structure
-
-        // Extension
-
-        if ( !(filename.endsWith(ModelBundle.TIO_BUNDLE_EXTENSION) || filename.endsWith(ModelBundle.TF_BUNDLE_EXTENSION)) ) {
-            throw new ValidatorException("No model bundle found with .tiobundle or .tfbundle extension at provided path");
-        }
-
-        // model.json
-
-        try {
-            if (!Arrays.asList(context.getResources().getAssets().list(filename)).contains(ModelBundle.TFMODEL_INFO_FILE)) {
-                throw new ValidatorException("No model.json at found in provided bundle");
-            }
-        } catch (Exception e) {
-            throw new ValidatorException("No model.json at found in provided bundle");
-        }
-
-        // Validate if JSON Can Be Read
-
-        String json = loadAssetJson();
-        JSONObject jsonObject = null;
-        JsonNode jsonNode = null;
-
-        if (json == null) {
-            throw new ValidatorException("Unable to read model.json");
-        }
-
-        try {
-            jsonObject = new JSONObject(json);
-            jsonNode = JsonLoader.fromString(json);
-        } catch (JSONException | IOException e) {
-            throw new ValidatorException("Unable to load model.json", e);
-        }
-
-        // Acquire Backend
-
-        String backend = "TFLite";
-
-        // Validate JSON with Schema
-
-        JsonSchema schema = jsonSchemaForBackend(backend);
-
-        if (schema == null) {
-            throw new ValidatorException("Unable to acquire model.json schema");
-        }
-
-        try {
-            ProcessingReport report = schema.validate(jsonNode);
-            if (!report.isSuccess()) {
-                throw new ValidatorException("model.json validation failed");
-            }
-        } catch (ProcessingException e) {
-            throw new ValidatorException("Unknown error encounted during model.json validation", e);
-        }
-
-        // Validate Assets
-
-        try {
-            validateAssetModelAssets(jsonObject);
-        } catch (Exception e) {
-            throw new ValidatorException("Unable to locate model file in bundle or other named assets");
-        }
-
-        // Custom Validator
-
-        if (customValidator != null && !customValidator.apply(filename, jsonObject)) {
-            throw new ValidatorException("Custom validator failed");
-        }
-    }
-
-    /** Validates a model bundle at a fully qualified File location */
-
-    private void validateFile(@Nullable BiFunction<String, JSONObject, Boolean> customValidator) throws ValidatorException {
-        // Validate Path
-
-        if (!file.exists()) {
-            throw new ValidatorException("No model bundle at found at provided path");
-        }
-
-        // Validate Bundle Structure
-
-        // Extension
-
-        if ( !(file.getPath().endsWith(ModelBundle.TIO_BUNDLE_EXTENSION) || file.getPath().endsWith(ModelBundle.TF_BUNDLE_EXTENSION)) ) {
-            throw new ValidatorException("No model bundle found with .tiobundle or .tfbundle extension at provided path");
-        }
-
-        // model.json
-
-        if (!jsonFile.exists()) {
-            throw new ValidatorException("No model.json at found in provided bundle");
-        }
-
-        // Validate if JSON Can Be Read
-
-        String json = loadFileJson();
-        JSONObject jsonObject = null;
-        JsonNode jsonNode = null;
-
-        if (json == null) {
-            throw new ValidatorException("Unable to read model.json");
-        }
-
-        try {
-            jsonObject = new JSONObject(json);
-            jsonNode = JsonLoader.fromString(json);
-        } catch (JSONException | IOException e) {
-            throw new ValidatorException("Unable to load model.json", e);
-        }
-
-        // Acquire Backend
-
-        String backend = "TFLite";
-
-        // Validate JSON with Schema
-
-        JsonSchema schema = jsonSchemaForBackend(backend);
-
-        if (schema == null) {
-            throw new ValidatorException("Unable to acquire model.json schema");
-        }
-
-        try {
-            ProcessingReport report = schema.validate(jsonNode);
-            if (!report.isSuccess()) {
-                throw new ValidatorException("model.json validation failed");
-            }
-        } catch (ProcessingException e) {
-            throw new ValidatorException("Unknown error encounted during model.json validation", e);
-        }
-
-        // Validate Assets
-
-        try {
-            validateFileModelAssets(jsonObject);
-        } catch (Exception e) {
-            throw new ValidatorException("Unable to locate model file in bundle or other named assets");
-        }
-
-        // Custom Validator
-
-        if (customValidator != null && !customValidator.apply(filename, jsonObject)) {
-            throw new ValidatorException("Custom validator failed");
-        }
-    }
-
-    /** Validates presence of model file if not placeholder model and any other model assets for a context.asset source */
-
-    private void validateAssetModelAssets(JSONObject jsonObject) throws JSONException, Exception {
-        if ( !(jsonObject.has("placeholder") && jsonObject.getBoolean("placeholder")) ) {
-            String modelFilename = jsonObject.getJSONObject("model").getString("file");
-            if (!Arrays.asList(context.getResources().getAssets().list(filename)).contains(modelFilename)) {
-                throw new Exception();
-            }
-        }
-
-        JSONArray outputs = jsonObject.getJSONArray("outputs");
-
-        for (int i = 0; i < outputs.length(); i++) {
-            JSONObject output = outputs.getJSONObject(i);
-            if (!output.has("labels")) {
-                continue;
-            }
-
-            String labelsFilename = output.getString("labels");
-            String assetPath = filename + "/" + ModelBundle.TFMODEL_ASSETS_DIRECTORY;
-
-            if (!Arrays.asList(context.getResources().getAssets().list(assetPath)).contains(labelsFilename)) {
-                throw new Exception();
-            }
-        }
-    }
-
-    /** Validates presence of model file if not placeholder model and any other model assets for a context.asset source */
-
-    private void validateFileModelAssets(JSONObject jsonObject) throws JSONException, Exception {
-        if ( !(jsonObject.has("placeholder") && jsonObject.getBoolean("placeholder")) ) {
-            String modelFilename = jsonObject.getJSONObject("model").getString("file");
-            File model = new File(file, modelFilename);
-
-            if (!model.exists()) {
-                throw new Exception();
-            }
-        }
-
-        JSONArray outputs = jsonObject.getJSONArray("outputs");
-
-        for (int i = 0; i < outputs.length(); i++) {
-            JSONObject output = outputs.getJSONObject(i);
-            if (!output.has("labels")) {
-                continue;
-            }
-
-            String labelsFilename = output.getString("labels");
-            File assetsDir = new File(file, ModelBundle.TFMODEL_ASSETS_DIRECTORY);
-            File asset = new File(assetsDir, labelsFilename);
-
-            if (!asset.exists()) {
-                throw new Exception();
-            }
-        }
-    }
-
-    /** Loads the model JSON from an context.asset source and returns null if unable */
-
-    private @Nullable String loadAssetJson() {
-        try {
-            return AndroidAssets.readTextFile(context, jsonFilename);
-        } catch (IOException e) {
-            return null;
-        }
-    }
-
-    /** Loads the model JSON from a file source and returns null if unable */
-
-    private @Nullable String loadFileJson() {
-        try {
-            return FileIO.readTextFile(jsonFile);
-        } catch (IOException e) {
-            return null;
-        }
-    }
+    public abstract void validate(@Nullable BiFunction<String, JSONObject, Boolean> customValidator) throws ValidatorException;
 
     /** Loads the JSON schema and returns null if unable */
 
-    private @Nullable JsonSchema jsonSchemaForBackend(@NonNull String backend){
+    protected @Nullable JsonSchema jsonSchemaForBackend(@NonNull String backend){
         try {
             String modelSchema = AndroidAssets.readTextFile(context, backend + "/model-schema.json");
             JsonNode schemaNode = JsonLoader.fromString(modelSchema);
